requirements.txt
tensorflow==2.11.0
keras==2.11.0
imageio==2.31.1
numpy==1.23.4
Pillow==9.3.0
scikit-image==0.21.0
scikit-learn==1.2.0
scipy==1.10.0